requirements.txt
attrs==22.2.0
autoflake==2.0.0
bandit==1.7.4
black==22.12.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
distlib==0.3.6
exceptiongroup==1.1.0
filelock==3.9.0
flake8==6.0.0
gitdb==4.0.10
GitPython==3.1.30
identify==2.5.12
idna==3.4
iniconfig==2.0.0
isort==5.11.4
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==22.0
pathspec==0.10.3
pbr==5.11.0
platformdirs==2.6.2
pluggy==1.0.0
pre-commit==2.21.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.2.0
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
smmap==5.0.0
stevedore==4.1.1
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
virtualenv==20.17.1
